+++ sync-for-expenses/src/sdk/models/operations/createconnection.ts
@@ -56,5 +56,5 @@ export class CreateConnectionResponse extends SpeakeasyBase {
      * Raw HTTP response; suitable for custom response parsing
      */
     @SpeakeasyMetadata()
-    rawResponse?: AxiosResponse;
+    rawResponse: AxiosResponse;
 }