+++ sync-for-expenses/src/sdk/models/operations/createsupplier.ts
@@ -64,5 +64,5 @@ export class CreateSupplierResponse extends SpeakeasyBase {
      * Raw HTTP response; suitable for custom response parsing
      */
     @SpeakeasyMetadata()
-    rawResponse?: AxiosResponse;
+    rawResponse: AxiosResponse;
 }