+++ sync-for-expenses/src/sdk/models/operations/initiatesync.ts
@@ -40,7 +40,7 @@ export class InitiateSyncResponse extends SpeakeasyBase {
      * Raw HTTP response; suitable for custom response parsing
      */
     @SpeakeasyMetadata()
-    rawResponse?: AxiosResponse;
+    rawResponse: AxiosResponse;
 
     /**
      * Returns the newly created SyncId