+++ sync-for-expenses/src/sdk/models/operations/updateexpensetransaction.ts
@@ -46,7 +46,7 @@ export class UpdateExpenseTransactionResponse extends SpeakeasyBase {
      * Raw HTTP response; suitable for custom response parsing
      */
     @SpeakeasyMetadata()
-    rawResponse?: AxiosResponse;
+    rawResponse: AxiosResponse;
 
     /**
      * Accepted