+++ sync-for-expenses/src/sdk/models/operations/updatesupplier.ts
@@ -70,7 +70,7 @@ export class UpdateSupplierResponse extends SpeakeasyBase {
      * Raw HTTP response; suitable for custom response parsing
      */
     @SpeakeasyMetadata()
-    rawResponse?: AxiosResponse;
+    rawResponse: AxiosResponse;
 
     /**
      * Success