+++ sync-for-expenses/src/sdk/models/operations/uploadexpenseattachment.ts
@@ -68,5 +68,5 @@ export class UploadExpenseAttachmentResponse extends SpeakeasyBase {
      * Raw HTTP response; suitable for custom response parsing
      */
     @SpeakeasyMetadata()
-    rawResponse?: AxiosResponse;
+    rawResponse: AxiosResponse;
 }